S-14f extract updates+branch-drift to web/modules/updates.js
Moves loadUpdates, renderUpdatesBadge, wireBranchDriftBanner, branch-drift
constants/helpers, pollUpdatesBadge, runUpdate, pollUpdateOutcome,
renderDiagnoseOffer, runDiagnose (~320 lines) out of app.js.
escapeHtmlUpdates is private to the module (only used by updates section).
wireBranchDriftBanner exported: DI-d into initSettings + called at boot.
initUpdates() wires check/apply buttons and starts the 5-min badge poll.
Build + 2887/2887 vitest green + headless smoke: 0 pageerror, 0 404s,
updatesPage visible, summary rendered, branch-drift banner+notice FOUND,
diagnose offer FOUND.
